Easy and medium puzzles now functional w/ win detection
Got the totals boxes working with user input and added (simple) win detection for easy and medium. My bad for some long chunks of code but some of the indexes weren't working as I intended so it was easier for me to just hardcode the indexes for the right input boxes. I'll come back and clean it up after adding some more functionality.